Rename the source code directory and create a `_utils` module
