+++ bh_testing/objects.py
@@ -2,30 +2,8 @@ import numpy as np
 import moderngl
 from PIL import Image
 
-DEGREES = np.pi / 180.0
+from ._utils import rotate_y, rotate_z, to_PIL_image
 
-def rotate_y(angle):
-    return np.array([
-        [np.cos(angle), 0, np.sin(angle)],
-        [0, 1, 0],
-        [-np.sin(angle), 0, np.cos(angle)]
-    ])
-
-def rotate_z(angle):
-    return np.array([
-        [np.cos(angle), -np.sin(angle), 0],
-        [np.sin(angle), np.cos(angle), 0],
-        [0, 0, 1]
-    ])
-
-def to_PIL_image(output_texture, size):
-    data = output_texture.read()
-    image = np.frombuffer(data, dtype=np.float32).reshape((size[1], size[0], 4))
-
-    # Convert to 8-bit per channel and create a PIL image
-    total_image = np.clip(image * 255, 0, 255).astype(np.uint8)
-    pil_image = Image.fromarray(total_image)
-    return pil_image
 
 class Camera:
     """
@@ -148,7 +126,7 @@ class Scene:
         self.black_hole = black_hole
         self.disk = disk
         self.background_texture = background_texture
-        self.show_disk = not disk is None
+        self.show_disk = disk is not None
 
         # Initiate
         self.ctx = moderngl.create_context(standalone=True)
